+++ controllers/AdminController.php
@@ -3,6 +3,7 @@
 namespace acmeCorp\humhub\modules\exampleBasic\controllers;
 
 use humhub\modules\admin\components\Controller;
+use Yii;
 
 class AdminController extends Controller
 {
@@ -14,6 +15,20 @@ class AdminController extends Controller
     public function actionIndex()
     {
         return $this->render('index');
+    }
+
+    public function actionConfig()
+    {
+        $model = $this->module->getConfiguration();
+
+        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            $this->view->saved();
+            return $this->redirect(['config']);
+        }
+
+        return $this->render('config', [
+            'model' => $model,
+        ]);
     }
 
 }