+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 - Fix #7: Update module resources path
 - Enh #12: Migration to Bootstrap 5 for HumHub 1.18
 - Enh #14: Improved Module Test GitHub Actions
+- Enh: Added Config Dialog
 
 1.0.2 (Unreleased)
 ----------------------